code/requirements.txt
PyYAML
pytest
pyhamcrest
requests